Modularized code by adding new folders and moving code into respective folders
