+++ floor-plan/src/app/interfaces/FloorPlanDocument.ts
